Rehooked up the importer to library model. Games will now import, but it still needs work.

+++ frontend/cpp/library/model/librarymodel.cpp
@@ -471,7 +471,7 @@ void LibraryModel::handleInsertGame( const GameData importData ) {
 }
 */
 
-void LibraryModel::setProgress( const qreal progress ) {
+void LibraryModel::setProgress( const int progress ) {
     if( progress == qmlProgress ) {
         return;
     }
